+++ CHANGELOG.md
@@ -14,16 +14,24 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 
 ### Adds
 
+- Adds the `DatePicker` component.
 - Adds new `Card` component.
 - Adds `spacing` prop to `Form` component to allow for spacing variations.
 - Integrates Chakra-UI into the codebase.
 - Exports Chakra's `Box`, `Center`, `Circle`, `Square`, `Stack`, `HStack`, and `VStack` components.
 - Composes the DS `Button` component from Chakra's `Button` component.
+- Composes the DS `Tabs` component from Chakra's `Tabs`, `TabList`, `Tab`, `TabPanels`, and `TabPanel` components.
+
+### Removes
+
+- Removes the `color-classes` SCSS mixin.
 
 ### Changes
 
-- Exports `VideoPlayerAspectRatios`
+- Exports `VideoPlayerAspectRatios`.
+- Adds additional props for the `DatePicker` component based on its TAD Addendum section.
 - Adds Tugboat QA configuration for PR preview deployments to replace Netlify.
+- Moved SCSS and CSS variable styles into a custom NYPL theme to extend Chakra's base theme.
 - Updates the `DatePicker`, `TextInput`, `Select`, and `Label` components to pass an optional `showOptReqLabel` prop to conditionally render "Required"/"Optional" in the label text.
 
 ### Fixes